refactor(llm-analytics): Decompose TextViewDisplay into modular components
Refactored the 984-line TextViewDisplay component into a modular architecture to improve maintainability and reduce code complexity. The component had deep nesting (8 levels) and mixed multiple responsibilities including parsing, rendering, and DOM manipulation.

Changes:
- Extracted parsing logic into dedicated utilities (segmentParser, urlParser, textHelpers)
- Created separate rendering components (SegmentRenderer, NestedContentRenderer, TextWithLinks, LineWithNumber)
- Replaced manual DOM manipulation (getElementById, classList) with React refs and useEffect hooks
- Implemented line number padding calculation for consistent visual alignment across all content
- Main component reduced from 984 lines to 199 lines
- Maximum nesting depth reduced from 8 to 3-4 levels

All existing functionality preserved including expandable sections, URL/event link handling, and line highlighting.
