+++ products/llm_analytics/frontend/text-view/components/LineWithNumber.tsx
@@ -1,0 +1,54 @@
+/**
+ * Line rendering component with line number and permalink functionality
+ * Uses ref for proper scroll handling
+ */
+import { useEffect, useRef } from 'react'
+
+import { Tooltip } from '@posthog/lemon-ui'
+
+import { copyToClipboard } from 'lib/utils/copyToClipboard'
+
+interface LineWithNumberProps {
+    lineNumber: number
+    content: string
+    isActive: boolean
+    padding: number
+}
+
+export function LineWithNumber({ lineNumber, content, isActive, padding }: LineWithNumberProps): JSX.Element {
+    const lineRef = useRef<HTMLSpanElement>(null)
+
+    useEffect(() => {
+        if (isActive && lineRef.current) {
+            lineRef.current.scrollIntoView({ behavior: 'smooth', block: 'center' })
+            lineRef.current.classList.add('bg-warning-highlight', 'border-l-4', 'border-warning')
+            const timer = setTimeout(() => {
+                lineRef.current?.classList.remove('bg-warning-highlight', 'border-l-4', 'border-warning')
+            }, 3000)
+            return () => clearTimeout(timer)
+        }
+    }, [isActive])
+
+    const handleCopyPermalink = (): void => {
+        const url = new URL(window.location.href)
+        url.searchParams.set('line', lineNumber.toString())
+        copyToClipboard(url.toString(), 'permalink')
+    }
+
+    const paddedLineNumber = padding > 0 ? lineNumber.toString().padStart(padding, '0') : lineNumber.toString()
+
+    return (
+        <span ref={lineRef}>
+            <Tooltip title="Click to copy permalink to this line">
+                <button
+                    type="button"
+                    className="text-muted hover:text-link cursor-pointer"
+                    onClick={handleCopyPermalink}
+                >
+                    L{paddedLineNumber}:
+                </button>
+            </Tooltip>
+            {content}
+        </span>
+    )
+}
